requirements.txt
alembic==1.9.1
amqp==5.1.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Login==0.6.2
Flask-Migrate==4.0.1
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
gunicorn==20.1.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.1
prompt-toolkit==3.0.36
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.7
six==1.16.0
SQLAlchemy==1.4.46
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
